+++ node/resolvers/shipping/index.ts
@@ -1,0 +1,9 @@
+import {
+  estimateShippingMutation,
+  selectDeliveryOptionMutation,
+} from './mutations'
+
+export const mutations = {
+  estimateShipping: estimateShippingMutation,
+  selectDeliveryOption: selectDeliveryOptionMutation,
+}